chore(db): chg decimal to float

+++ database/migrations/2022_11_23_193950_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('brand')->nullable();
             $table->string('name');
             $table->string('tags_json')->nullable();
-            $table->decimal('default_price');
+            $table->float('default_price');
             $table->integer('low_qty')->nullable();
             $table->string('unit', 10);
             $table->timestamps();
